Ship unparseable FLOAT values verbatim from the add-component coercer
